# Optimize topic message lookup migration