**Description**:
<!--
One or two line summary of what this PR does and why it is needed, followed by a list
of changes in imperative, present tense for use in the commit message or changelog. Example:

This PR modifies ... in order to support ...
* Add config property
* Change column name
* Remove ...
-->
This PR optimizes the topic message lookup migration.

- Add functionality to group topic ids by citus shard
- Add functionality to find top topics with more than 1 million messages in a partition
- Insert topic message lookups for non-top topics in batches by shard for each partition
- Insert topic message lookups for top topics one at a time for each partition
- Fix slow query to determine whether to skip partition

**Related issue(s)**:

Fixes #8512 

**Notes for reviewer**:
<!-- Provide logs, performance numbers or screenshots of the new functionality -->
The previous optimized version with a fixed 12 top topics from mainnet db took about 52 minutes to complete.

The enhanced version in this PR is supposed to take less time, however I just ran it for the last several partitions:

```
2024-06-24T20:43:52.415Z  INFO main c.h.m.i.m.TopicMessageLookupMigration Migrated topic_message_lookup with 10159 rows for partition topic_message_p2024_06 in 22.56 s
2024-06-24T20:43:52.485Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Starting asynchronous migration
2024-06-24T20:43:52.485Z  INFO main c.h.m.i.m.TopicMessageLookupMigration Ran migration Backfill topic_message_lookup table in 28.17 s.
2024-06-24T20:44:31.221Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Migrated topic_message_lookup with 15408 rows for partition topic_message_p2024_05 in 37.15 s
2024-06-24T20:44:58.636Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Migrated topic_message_lookup with 34514 rows for partition topic_message_p2024_04 in 27.38 s
2024-06-24T20:44:58.677Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Completed iteration 2 with last value: topic_message_p2024_03
2024-06-24T20:45:43.501Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Migrated topic_message_lookup with 67023 rows for partition topic_message_p2024_03 in 44.82 s
2024-06-24T20:46:47.627Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Migrated topic_message_lookup with 4825 rows for partition topic_message_p2024_02 in 1.068 min
2024-06-24T20:46:47.668Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Completed iteration 4 with last value: topic_message_p2024_01
2024-06-24T20:47:43.565Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Migrated topic_message_lookup with 2846 rows for partition topic_message_p2024_01 in 55.90 s
2024-06-24T20:47:43.602Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Completed iteration 5 with last value: topic_message_p2023_12
2024-06-24T20:48:39.752Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Migrated topic_message_lookup with 1122 rows for partition topic_message_p2023_12 in 56.15 s
2024-06-24T20:48:39.792Z  INFO single-1 c.h.m.i.m.TopicMessageLookupMigration Completed iteration 6 with last value: topic_message_p2023_11
```

**Checklist**

- [ ] Documented (Code comments, README, etc.)
- [x] Tested (unit, integration, etc.)